Migrate terra-modal-manager docs and update dependencies.
