+++ packages/terra-abstract-modal/package.json
@@ -38,7 +38,7 @@
     "wicg-inert": "^3.0.0"
   },
   "devDependencies": {
-    "terra-icon": "^3.1.0"
+    "terra-icon": "^3.19.0"
   },
   "scripts": {
     "compile": "babel --root-mode upward src --out-dir lib --copy-files",
