+++ packages/terra-application-layout/package.json
@@ -37,8 +37,8 @@
     "terra-application-menu-layout": "^3.30.0",
     "terra-application-name": "^3.39.0",
     "terra-application-utility": "^2.44.0",
-    "terra-breakpoints": "^2.0.0",
-    "terra-icon": "^3.1.0",
+    "terra-breakpoints": "^2.19.0",
+    "terra-icon": "^3.19.0",
     "terra-modal-manager": "^6.54.0",
     "terra-navigation-layout": "^5.33.1",
     "terra-navigation-side-menu": "^2.39.0",
