+++ packages/terra-application-links/package.json
@@ -33,7 +33,7 @@
     "keycode-js": "^3.1.0",
     "prop-types": "^15.5.8",
     "resize-observer-polyfill": "^1.4.1",
-    "terra-icon": "^3.1.0",
+    "terra-icon": "^3.19.0",
     "terra-popup": "^6.54.0",
     "terra-theme-context": "^1.7.0"
   },
