+++ packages/terra-application-navigation/package.json
@@ -43,10 +43,10 @@
     "terra-action-footer": "^2.0.0",
     "terra-action-header": "^2.0.0",
     "terra-avatar": "^3.8.0",
-    "terra-breakpoints": "^2.0.0",
+    "terra-breakpoints": "^2.19.0",
     "terra-button": "^3.3.0",
     "terra-content-container": "^3.0.0",
-    "terra-icon": "^3.11.0",
+    "terra-icon": "^3.19.0",
     "terra-image": "^3.0.0",
     "terra-overlay": "^3.0.0",
     "terra-popup": "^6.54.0",
