+++ packages/terra-application-utility/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated `terra-icon` from `v3.1.0` to `v3.19.0`.
+
 ## 2.44.0 - (June 8, 2021)
 
 * Changed
