+++ packages/terra-collapsible-menu-view/package.json
@@ -32,14 +32,14 @@
     "resize-observer-polyfill": "^1.4.1",
     "terra-button": "^3.3.0",
     "terra-button-group": "^3.0.0",
-    "terra-form-checkbox": "^4.0.0",
+    "terra-form-checkbox": "^4.8.0",
     "terra-list": "^4.0.0",
     "terra-menu": "^6.54.0",
     "terra-mixins": "^1.0.0",
     "terra-theme-context": "^1.7.0"
   },
   "devDependencies": {
-    "terra-icon": "^3.1.0"
+    "terra-icon": "^3.19.0"
   },
   "scripts": {
     "compile": "babel --root-mode upward src --out-dir lib --copy-files",
