+++ packages/terra-date-input/package.json
@@ -32,7 +32,7 @@
     "keycode-js": "^3.1.0",
     "prop-types": "^15.5.8",
     "terra-form-input": "^4.4.0",
-    "terra-icon": "^3.1.0",
+    "terra-icon": "^3.19.0",
     "terra-mixins": "^1.0.0",
     "terra-theme-context": "^1.7.0",
     "terra-visually-hidden-text": "^2.0.0",
