+++ packages/terra-date-picker/package.json
@@ -34,11 +34,11 @@
     "prop-types": "^15.5.8",
     "react-onclickoutside": "^6.7.1",
     "react-portal": "^4.1.2",
-    "terra-breakpoints": "^2.0.0",
+    "terra-breakpoints": "^2.19.0",
     "terra-button": "^3.3.0",
-    "terra-form-field": "^3.1.0",
+    "terra-form-field": "^4.5.0",
     "terra-form-input": "^4.4.0",
-    "terra-icon": "^3.1.0",
+    "terra-icon": "^3.19.0",
     "terra-popup": "^6.54.0",
     "terra-responsive-element": "^5.0.0",
     "terra-theme-context": "^1.7.0",
