+++ packages/terra-date-time-picker/package.json
@@ -38,7 +38,7 @@
     "terra-time-input": "^4.41.0"
   },
   "devDependencies": {
-    "terra-form-field": "^3.1.0",
+    "terra-form-field": "^4.5.0",
     "terra-modal-manager": "^6.54.0"
   },
   "scripts": {
