+++ packages/terra-form-validation/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated `terra-form-input` from `v3.7.0` to `v4.4.0`.
+
 ## 1.62.0 - (July 20, 2021)
 
 * Changed
