+++ packages/terra-form-validation/package.json
@@ -37,7 +37,7 @@
     "terra-date-picker": "^4.64.0",
     "terra-form-checkbox": "^4.8.0",
     "terra-form-field": "^4.5.0",
-    "terra-form-input": "^3.7.0",
+    "terra-form-input": "^4.4.0",
     "terra-form-radio": "^4.12.0",
     "terra-form-select": "^6.8.0",
     "terra-form-textarea": "^5.1.0",
