+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Updated
+  * Migrated terra-modal-manager docs
+
 ## 1.0.0 - (July 20, 2021)
 
 * Initial stable release
