+++ packages/terra-framework-docs/package.json
@@ -43,10 +43,12 @@
     "terra-button": "^3.3.0",
     "terra-content-container": "^3.0.0",
     "terra-disclosure-manager": "^4.41.0",
-    "terra-form-input": "^3.7.0",
+    "terra-form-input": "^4.4.0",
     "terra-grid": "^6.0.0",
     "terra-modal-manager": "^6.54.0",
     "terra-section-header": "^2.0.0",
-    "terra-slide-panel-manager": "^5.55.0"
+    "terra-slide-panel-manager": "^5.55.0",
+    "terra-collapsible-menu-view": "^6.54.0",
+    "terra-form-select": "^6.8.0"
   }
 }
