+++ packages/terra-framework-docs/src/terra-dev-site/doc/modal-manager/About.1.doc.mdx
@@ -1,6 +1,6 @@
 import { Badge } from 'terra-modal-manager/package.json?dev-site-package';
 
-import ModalManagerExample from '../example/ModalManagerExample?dev-site-example';
+import ModalManagerExample from './example/ModalManagerExample?dev-site-example';
 
 import ModalManagerPropsTable from 'terra-modal-manager/src/ModalManager?dev-site-props-table';
