+++ packages/terra-framework-docs/src/terra-dev-site/test/modal-manager/ModalManagerDefault.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import DisclosureComponent from './DisclosureIntegrationComponent';
-import ModalManager from '../../../ModalManager';
+import ModalManager from 'terra-modal-manager';
+import DisclosureComponent from './DisclosureComponent';
 import styles from './ModalManagerCommon.test.module.scss';
 
 const cx = classNames.bind(styles);
