+++ packages/terra-framework-docs/src/terra-dev-site/test/modal-manager/ModalManagerWithDisclosureAccessory.test.jsx
@@ -1,26 +1,19 @@
 import React from 'react';
 import classNames from 'classnames/bind';
+import ModalManager from 'terra-modal-manager';
 import DisclosureComponent from './DisclosureComponent';
-import ModalManager from '../../../ModalManager';
 import styles from './ModalManagerCommon.test.module.scss';
 
 const cx = classNames.bind(styles);
 
-const DisclosureContainer = (wrappedContent) => (
-  <div data-disclosure-container className={cx('disclosure-container')}>
-    I am wrapping the content with a purple background!
-    {wrappedContent}
-  </div>
-);
-
-const ModalManagerWithDisclosureContainer = () => (
+const ModalManagerWithDisclosureAccessory = () => (
   <div role="main" className={cx('content-wrapper')}>
     <ModalManager
-      withDisclosureContainer={DisclosureContainer}
+      disclosureAccessory={<div className={cx('disclosure-accessory')}>Disclosure Accessory</div>}
     >
       <DisclosureComponent identifier="root-component" disclosureType="modal" renderHeaderAdapter />
     </ModalManager>
   </div>
 );
 
-export default ModalManagerWithDisclosureContainer;
+export default ModalManagerWithDisclosureAccessory;
