+++ packages/terra-hookshot/package.json
@@ -39,7 +39,7 @@
   },
   "devDependencies": {
     "terra-button": "^3.3.0",
-    "terra-form-input": "^3.7.0"
+    "terra-form-input": "^4.4.0"
   },
   "scripts": {
     "compile": "babel --root-mode upward src --out-dir lib --copy-files",
