+++ packages/terra-layout/package.json
@@ -39,7 +39,7 @@
   "devDependencies": {
     "react-router-dom": "^5.0.0",
     "terra-button": "^3.3.0",
-    "terra-icon": "^3.1.0"
+    "terra-icon": "^3.19.0"
   },
   "scripts": {
     "compile": "babel --root-mode upward src --out-dir lib --copy-files",
