+++ packages/terra-modal-manager/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Moved docs to terra-framework-docs
+
 ## 6.54.0 - (July 13, 2021)
 
 * Changed
