+++ packages/terra-modal-manager/package.json
@@ -35,10 +35,6 @@
     "terra-slide-group": "^4.27.0",
     "terra-theme-context": "^1.7.0"
   },
-  "devDependencies": {
-    "terra-collapsible-menu-view": "^6.54.0",
-    "terra-form-select": "^6.8.0"
-  },
   "scripts": {
     "compile": "babel --root-mode upward src --out-dir lib --copy-files",
     "lint": "npm run lint:js && npm run lint:scss",
