+++ packages/terra-navigation-layout/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated `terra-icon` from `v3.1.0` to `v3.19.0`.
+
 ## 5.33.1 - (June 8, 2021)
 
 * Changed
