+++ packages/terra-navigation-side-menu/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated `terra-icon` from `v3.1.0` to `v3.19.0`.
+
 ## 2.39.0 - (June 8, 2021)
 
 * Changed
