+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -2,10 +2,13 @@
 
 ## Unreleased
 
+* Changed
+  * Updated `terra-icon` from `v3.1.0` to `v3.19.0`.
+
 ## 4.16.0 - (July 20, 2021)
 
 * Changed
-  * Updated the upgrade guide to state to replace `secondaryAction` with `rejectAction` instead of `primaryAction` 
+  * Updated the upgrade guide to state to replace `secondaryAction` with `rejectAction` instead of `primaryAction`
 
 ## 4.15.0 - (July 13, 2021)
