+++ packages/terra-notification-dialog/package.json
@@ -40,7 +40,7 @@
   "devDependencies": {
     "terra-breakpoints": "^2.19.0",
     "terra-disclosure-manager": "^4.41.0",
-    "terra-icon": "^3.1.0",
+    "terra-icon": "^3.19.0",
     "terra-modal-manager": "^6.54.0",
     "terra-popup": "^6.54.0"
   },
